refactor screenshot editor into multiple files
